+++ insights-ui/src/app/public-equities/tickers/[tickerKey]/page.tsx
@@ -1,13 +1,7 @@
 import Breadcrumbs from '@/components/ui/Breadcrumbs';
 import RadarChart from '@/components/ui/RadarChart';
-import {
-  CriteriaEvaluation,
-  IndustryGroupCriteria,
-  PerformanceChecklistItem,
-  SpiderGraphForTicker,
-  SpiderGraphPie,
-  TickerReport,
-} from '@/types/public-equity/ticker-report';
+import { IndustryGroupCriteria } from '@/types/public-equity/criteria-types';
+import { CriteriaEvaluation, PerformanceChecklistItem, SpiderGraphForTicker, SpiderGraphPie, TickerReport } from '@/types/public-equity/ticker-report';
 import { getReportName } from '@/util/report-utils';
 import { BreadcrumbsOjbect } from '@dodao/web-core/components/core/breadcrumbs/BreadcrumbsWithChevrons';
 import PageWrapper from '@dodao/web-core/components/core/page/PageWrapper';
